src/requirements.txt
fasttext==0.9.2
huggingface_hub==0.35.3
iso639-lang==2.6.3
